# Improve predictability of writeUtf8/writeAscii performance