… buffer/src/test/java/io/netty/buffer/ByteBufUtilTest.java
… microbench/src/main/java/io/netty/microbench/buffer/Utf8EncodingBenchmark.java
… microbench/src/main/resources/Utf8Samples.txt