@franz1981 2020-06-23
Need to explain the context of this PR with https://github.com/eclipse-vertx/vert.x/pull/3462#issuecomment-647038310: when user code mix different `ByteBuf` subtypes or just has it's own stack depth, the `-XX:MaxInlineLevel` config (9 on jdk 8 and 15 on jdk 14) could prevent certain calls to be inlined causing unwanted bound checks, prevent vectorizations...hitting the stability of performance we see in the micro-benchmarks or just in an ideal world.

This duplication is aiming to makes the performance a bit more stable (sadly `PlatformDependent::putByte` has still some depth and `CharSequence:.charAt` too, but the benchmarks already shows some improvement.
I don't think I will left the benchmark changes to be there in the final PR because are too tied to how the JDK is configured.