@franz1981 2020-06-23
I've tweaked the benchs and impl to shows what's the effects of these changes...
this PR:
```
Benchmark                             (direct)   Mode  Cnt        Score        Error  Units
ByteBufUtilBenchmark.nestedWriteUtf8      true  thrpt   20  1135949.136 ±  15508.180  ops/s
ByteBufUtilBenchmark.nestedWriteUtf8     false  thrpt   20  1115468.247 ±  65425.276  ops/s
ByteBufUtilBenchmark.writeUtf8String      true  thrpt   20  8237112.811 ± 274131.929  ops/s
ByteBufUtilBenchmark.writeUtf8String     false  thrpt   20  4164310.459 ± 117297.047  ops/s
```
Before this PR with `-XX:MaxInlineLevel=9`:
```
Benchmark                             (direct)   Mode  Cnt        Score        Error  Units
ByteBufUtilBenchmark.nestedWriteUtf8      true  thrpt   20   738208.471 ±  13030.643  ops/s
ByteBufUtilBenchmark.nestedWriteUtf8     false  thrpt   20   815001.393 ±  48192.092  ops/s
ByteBufUtilBenchmark.writeUtf8String      true  thrpt   20  5859888.733 ± 143176.688  ops/s
ByteBufUtilBenchmark.writeUtf8String     false  thrpt   20  4484238.436 ±  64584.530  ops/s
```
Before this PR with `-XX:MaxInlineLevel=15`:
```
Benchmark                             (direct)   Mode  Cnt        Score       Error  Units
ByteBufUtilBenchmark.nestedWriteUtf8      true  thrpt   20  5499785.717 ± 59616.413  ops/s
ByteBufUtilBenchmark.nestedWriteUtf8     false  thrpt   20  4954847.853 ± 70535.926  ops/s
ByteBufUtilBenchmark.writeUtf8String      true  thrpt   20  5826841.982 ± 66342.052  ops/s
ByteBufUtilBenchmark.writeUtf8String     false  thrpt   20  4597271.515 ± 17988.082  ops/s
```
on jdk8 `nestedWriteUtf8` shows what's the possible effects of failing to inline some of the Netty `buffer` calls (or `PlatformDependent` too), while `String::charAt` isn't inlined (sadly, but could be fixed too on jdk 8 with some "hack").
`writeUtf8String` shows instead the effect of this PR while everything is correctly inlined instead, and performance are slightly improved too ;)
In this benchmark I haven't shows what's the effect of polymorphism (bi,tri, mega...) but I'm sure is needed, because just inlining shows the motivation behind the duplicated code.