@franz1981 2020-06-23
I've pushed a further commit that is manually inlining the slow path of `writeUtf8` ie `writeUtf8Surrogate` in order to make the methods > `-XX:FreqInlineSize` (ie 325 bytes)
eg `io.netty.buffer.ByteBufUtil::unsafeWriteUtf8 (358 bytes)   hot method too big` and 
`io.netty.buffer.ByteBufUtil::arrayWriteUtf8 (328 bytes)   hot method too big`
The results seems nice:
```
Benchmark                             (direct)   Mode  Cnt        Score        Error  Units
ByteBufUtilBenchmark.nestedWriteUtf8      true  thrpt   20  6524381.794 ± 102816.039  ops/s
ByteBufUtilBenchmark.nestedWriteUtf8     false  thrpt   20  4101033.546 ± 238837.837  ops/s
ByteBufUtilBenchmark.writeUtf8String      true  thrpt   20  7958679.895 ± 687717.569  ops/s
ByteBufUtilBenchmark.writeUtf8String     false  thrpt   20  4502402.790 ±  91698.297  ops/s
```
But I would like to see how the ASM looks like