@franz1981 2020-06-23
I'm not very used to achieve good numbers by playing so much with the JDK implementation details: I would like some good eyes to help me a bit on this (@nitsanw if he has the chance to look at this)