@franz1981 2020-06-23
In the last commit I've tried to use `PlatformDependent` on `byte[]` to save the array bound-checks to happen and indeed I see some improvement: it couldn't be used before becaure now the `xxxWriteUtf8` methods are not inlineable anymore...
```
Benchmark                             (direct)   Mode  Cnt        Score        Error  Units
ByteBufUtilBenchmark.nestedWriteUtf8      true  thrpt   20  7848114.042 ± 229818.691  ops/s
ByteBufUtilBenchmark.nestedWriteUtf8     false  thrpt   20  5907367.968 ± 321506.155  ops/s
ByteBufUtilBenchmark.writeUtf8String      true  thrpt   20  8264472.002 ±  49110.805  ops/s
ByteBufUtilBenchmark.writeUtf8String     false  thrpt   20  6166188.985 ± 336638.483  ops/s
```