@franz1981 2020-06-24
The thing that isn't 100% convincing me is the magnitude of error that doesn't seem proportional to the gain: I still need to investigate what's going on.