@nitsanw 2020-06-25
I know the code is based off of some of the hacks I did on this blog post : http://psy-lob-saw.blogspot.com/2012/12/encode-utf-8-string-to-bytebuffer-faster.html

But it's been a while :-), reading through the patch though, I recall having an implementation somewhere that used the same code for array and memory case, there should be no need for the code duplication. Also, the blog post had a better data set to use for the benchmark IIRC.