@njhill 2020-06-25
I've [commented in past](https://github.com/netty/netty/pull/9260#issuecomment-504594301) how sensitive `ByteBuf` access is in general to stack-depth due to transitive optimizations from inlining (eliminating accessibility checks). It means that in cases like this where there are many granular accesses, how deep the call stack happens to be makes an enormous difference to performance.

Having started to go down the rabbit hole of manually inlining parts of the `ByteBuf` hierarchy I came to the conclusion that it's not really worth trying to trim the stack depth by one or two calls for specific applications if that results in a lot of duplication and less-maintainable code, since it only helps in the borderline cases (the app in question might add an additional level of method indirection, at which point the benefit is lost).

Instead I feel we should just strongly advertise the fact that users can often get a big performance boost by setting `-XX:MaxInlineLevel` much higher than the default (at least for java < 14). I haven't encountered any cases where doing so is detrimental.

But otherwise I do really like the idea of speeding up reading/writing utf8 to `ByteBuf`s if/where possible given its prevalence on critical paths.

One specific question about the changes is why inlining `writeUtf8Surrogate` makes sense given how rarely that path would be hit?