@franz1981 2020-06-25
@njhill 
> One specific question about the changes is why inlining writeUtf8Surrogate makes sense given how rarely that path would be hit?

https://github.com/netty/netty/pull/10368#issuecomment-648217848

Tldr to save the big method from being inlined and "refill" the inlining budget: this would save from being dependent by the user code stack depth and would make any PlatformDependent call inlineable (because < 9th stack depth).
By consequence, CharSequence::chatAt became a bimorphic call for most usages (String or StringBuilder), getting inlined as well.